Add lock, and validate before and after the object deletion etc.

+++ main.go
@@ -10,13 +10,13 @@ import (
 
 func main() {
 	var (
-		numObj      int64
+		numObj      int
 		numWorker   int
 		sizePattern string
 		time        int64
 		bucketName  string
 	)
-	flag.Int64Var(&numObj, "num_obj", 10, "The number of objects.")
+	flag.IntVar(&numObj, "num_obj", 10, "The number of objects.")
 	flag.IntVar(&numWorker, "num_worker", 1, "The number of workers.")
 	flag.StringVar(&sizePattern, "size", "4k", "The size of object. Should be in the form like \"4k-2m\" or \"8k\".")
 	flag.Int64Var(&time, "time", 3, "Time duration in seconds to run the workload.")
